build(client/packages): transform current dir in the new monorepo; replace bun with pnpm
