+++ .gitignore
@@ -9,10 +9,11 @@ yarn-debug.log*
 yarn-error.log*
 lerna-debug.log*
 .pnpm-debug.log*
-
+pnpm-lock.yaml
 # Caches
 
 .cache
+.turbo
 
 # Diagnostic reports (https://nodejs.org/api/report.html)
