+++ README.md
@@ -4,16 +4,9 @@ Testing SolidJs concepts with new tanstack router
 ## Internal Notes
 
 - [X] Configure boilerplate
-- [ ] Add a new component
-- [ ] Add a new page
-- [ ] Add a new layout
-- [ ] Add a new hook
-- [ ] Add a new context
-- [ ] Add a new store
-- [ ] Add a new service
-- [ ] Add a new hook
-- [ ] Add a new context
-- [ ] Add a new store
-- [ ] Add a new service
+- [X] Configure Monorepo
+- [X] Create auth Pages
+- [ ] Create auth service with better-auth
+- [ ] More goals add later...
 
 [![Commitizen friendly](https://img.shields.io/badge/commitizen-friendly-brightgreen.svg)](http://commitizen.github.io/cz-cli/)
