+++ client/index.html
@@ -1,0 +1,8 @@
+<!doctype html>
+<html lang="en">
+  <head></head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
+  </body>
+</html>
