+++ client/src/assets/svgs/logo.tsx
