+++ client/src/lib/auth-client.ts
