+++ client/src/lib/auth.ts
