+++ client/src/routes/__root.tsx
