+++ client/src/routes/_private/_layout.hotspots/index.tsx
