+++ client/src/routes/_private/_layout.tsx
