+++ client/src/routes/register.tsx
