+++ package.json
@@ -1,60 +1,49 @@
 {
-  "name": "solid-tsrouter",
-  "version": "0.0.0",
-  "author": {
-    "name": "Yam P. Martins",
-    "email": "ymprado2010@gmail.com",
-    "url": "https://github.com/yammartins"
-  },
-  "module": "main.ts",
-  "type": "module",
+  "private": true,
   "scripts": {
-    "dev": "vite --port 3000",
-    "build": "vite build && tsc --noEmit",
-    "serve": "vite preview",
-    "start": "vite",
-    "check": "biome check .",
-    "check:fix": "biome check --fix .",
+    "dev": "turbo run dev",
+    "build": "turbo run build",
+    "serve": "turbo run serve",
+    "start": "turbo run start",
+    "check": "turbo run check",
+    "check:fix": "turbo run check:fix",
     "check:unused": "knip --max-issues 1",
     "check:unused:fix": "knip --fix",
     "commit": "cz",
-    "typecheck": "tsc --noEmit",
+    "typecheck": "turbo run typecheck",
     "knip": "knip"
   },
   "devDependencies": {
     "@biomejs/biome": "1.9.4",
     "@commitlint/cli": "19.6.1",
     "@commitlint/config-conventional": "19.6.0",
-    "@tanstack/router-plugin": "^1.111.3",
-    "@types/bun": "latest",
-    "@types/js-cookie": "^3.0.6",
-    "@types/node": "^22.13.5",
-    "clsx": "2.1.1",
     "commitizen": "4.3.1",
     "cz-conventional-changelog": "^3.3.0",
     "knip": "^5.44.5",
     "lefthook": "^1.11.0",
-    "tailwind-merge": "3.0.2",
-    "typescript": "^5.7.3",
-    "vite": "^6.1.1",
-    "vite-plugin-solid": "^2.11.2"
-  },
-  "peerDependencies": {
+    "turbo": "2.4.4",
     "typescript": "^5.7.3"
   },
-  "dependencies": {
-    "@modular-forms/solid": "^0.25.1",
-    "@prisma/client": "^6.4.1",
-    "@prisma/extension-accelerate": "^1.2.2",
-    "@solidjs/meta": "^0.29.4",
-    "@tailwindcss/postcss": "^4.0.8",
-    "@tailwindcss/vite": "^4.0.8",
-    "@tanstack/solid-router": "^1.111.3",
-    "better-auth": "^1.2.3",
-    "js-cookie": "^3.0.5",
-    "postcss": "^8.5.3",
-    "prisma": "^6.4.1",
-    "tailwindcss": "^4.0.8",
-    "zod": "^3.24.2"
+  "packageManager": "pnpm@10.6.1",
+  "pnpm": {
+    "ignoredBuiltDependencies": [
+      "@biomejs/biome",
+      "@prisma/client",
+      "@prisma/engines",
+      "esbuild",
+      "lefthook",
+      "prisma"
+    ],
+    "onlyBuiltDependencies": [
+      "@biomejs/biome",
+      "@prisma/client",
+      "@prisma/engines",
+      "esbuild",
+      "lefthook",
+      "prisma"
+    ]
+  },
+  "engines": {
+    "node": ">=20"
   }
 }
