+++ packages/biome-config/notes.md
@@ -1,0 +1,1 @@
+## NOT implemented yet
