+++ packages/biome-config/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@solid-tsrouter/biome-config",
+  "version": "0.0.0",
+  "license": "MIT",
+  "private": true
+}
