+++ packages/typescript-config/base.json
@@ -1,31 +1,24 @@
 {
+  "$schema": "https://json.schemastore.org/tsconfig",
   "compilerOptions": {
-    "lib": ["DOM", "DOM.Iterable", "ES2022"],
-    "target": "ESNext",
     "module": "ESNext",
+    "moduleResolution": "bundler",
     "moduleDetection": "force",
-    "jsx": "preserve",
-    "jsxImportSource": "solid-js",
     "allowJs": true,
 
-    "moduleResolution": "bundler",
     "allowImportingTsExtensions": true,
     "verbatimModuleSyntax": true,
     "noEmit": true,
 
+    "sourceMap": true,
     "strict": true,
     "skipLibCheck": true,
     "noFallthroughCasesInSwitch": true,
+    "forceConsistentCasingInFileNames": true,
 
+    "esModuleInterop": true,
     "noUnusedLocals": false,
     "noUnusedParameters": false,
-    "noPropertyAccessFromIndexSignature": false,
-
-    "baseUrl": ".",
-    "paths": {
-      "@/*": ["./src/*"]
-    }
-  },
-  "files": ["src/main.tsx"],
-  "include": ["src"]
+    "noPropertyAccessFromIndexSignature": false
+  }
 }
