+++ packages/typescript-config/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@solid-tsrouter/typescript-config",
+  "version": "1.0.0",
+  "license": "MIT",
+  "private": true
+}
